requirements.txt
# Core dependencies for Google Sheets script
pandas==2.1.3
google-auth==2.23.4
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Scrapy and related packages
Scrapy==2.11.0
Twisted==22.10.0  # Compatible with Scrapy 2.11.0
scrapy-playwright==0.0.33
playwright==1.40.0

# IPython for interactive development
ipython==8.17.2

# Other dependencies
asttokens==2.4.1
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.5
cssselect==1.2.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
greenlet==3.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jmespath==1.0.1
lxml==4.9.3
matplotlib-inline==0.1.6
packaging==23.2
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.41
Protego==0.3.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==11.0.1
Pygments==2.17.1
pyOpenSSL==23.3.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
service-identity==23.1.0
six==1.16.0
stack-data==0.6.3
tldextract==5.1.1
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.1.0
w3lib==2.1.2
wcwidth==0.2.10
zope.interface==6.1